+++ database/migrations/2025_10_29_111801_add_accomodation_on_quotation.php
@@ -13,10 +13,10 @@ return new class extends Migration
     {
         Schema::table('quotations', function (Blueprint $table) {
             $table->string('accommodation_wilayah')->nullable();
-            $table->integer('accommodation_hotel_rooms')->nullable();
-            $table->integer('accommodation_people')->nullable();
-            $table->integer('accommodation_target_days')->nullable();
-            $table->integer('accommodation_plane_ticket_price')->nullable();
+            $table->bigInteger('accommodation_hotel_rooms')->nullable();
+            $table->bigInteger('accommodation_people')->nullable();
+            $table->bigInteger('accommodation_target_days')->nullable();
+            $table->bigInteger('accommodation_plane_ticket_price')->nullable();
 
 
         });
